TD-1388: remove duplicated code
